Adding prints to see the number of points in each cloud and octomap Reading a quality file as an input alongside with the obj model

+++ nbv_simulation/View_Space.h
@@ -88,7 +88,7 @@ public:
      * Reading the SFM input file and loading the set of already used views
      * @param i The number of views already added
      */
-    double read_sfm_views(int i);
+    int read_sfm_views(int i);
 };
 
 void add_trajectory_to_cloud(Eigen::Matrix4d now_camera_pose_world,
